Added health system

+++ Assets/Scripts/Generic/Objects/WeaponObject.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class WeaponObject : BaseObject
 {
     public float range = 100.0f;
-    public float damagePerShot = 10.0f;
+    public int damagePerShot = 10;
     public int magazineSize = 30;
     public float reloadTime = 1.0f;
     public float TimeBetweenShots = 1.0f;
